Updates after feedback
Did some tidy up.

Removed the need to specify a localstoage key for the sample auth module. It uses a fixed key and uuid for uniqueness.

Moved the create app definition action into the developer action module.

Updates after feedbacks to tighten things.

Added logic in auth endpoint example that helps show the scheme (light/dark) specified in the role mapping.
