--- how-to/customize-workspace/client/src/modules/actions/app-definition/index.ts
@@ -1,6 +1,0 @@
-import type { ModuleImplementation, ModuleTypes } from "customize-workspace/shapes/module-shapes";
-import { AppDefinitionActions } from "./actions";
-
-export const entryPoints: { [type in ModuleTypes]?: ModuleImplementation } = {
-	actions: new AppDefinitionActions()
-};
