+++ how-to/customize-workspace/client/src/modules/auth/example/util.ts
@@ -1,24 +1,19 @@
 import type { ExampleUser } from "./shapes";
 
-export function getCurrentUser(userSessionId: string): ExampleUser {
-	if (userSessionId === undefined || userSessionId === null) {
-		return null;
-	}
-	const storedUser = localStorage.getItem(userSessionId);
+export const EXAMPLE_AUTH_CURRENT_USER_KEY = `${fin.me.identity.uuid}-EXAMPLE_AUTH_CURRENT_USER`;
+
+export function getCurrentUser(): ExampleUser | null {
+	const storedUser = localStorage.getItem(EXAMPLE_AUTH_CURRENT_USER_KEY);
 	if (storedUser === null) {
 		return null;
 	}
 	return JSON.parse(storedUser) as ExampleUser;
 }
 
-export function setCurrentUser(userSessionId: string, user: ExampleUser): boolean {
-	if (userSessionId === undefined || userSessionId === null) {
-		return false;
-	}
-	localStorage.setItem(userSessionId, JSON.stringify(user));
-	return true;
+export function setCurrentUser(user: ExampleUser): void {
+	localStorage.setItem(EXAMPLE_AUTH_CURRENT_USER_KEY, JSON.stringify(user));
 }
 
-export function clearCurrentUser(userSessionId: string) {
-	localStorage.removeItem(userSessionId);
+export function clearCurrentUser() {
+	localStorage.removeItem(EXAMPLE_AUTH_CURRENT_USER_KEY);
 }
