+++ how-to/customize-workspace/client/webpack.config.js
@@ -272,33 +272,6 @@ module.exports = [
 		}
 	},
 	{
-		entry: './client/src/modules/actions/app-definition/index.ts',
-		devtool: 'inline-source-map',
-		module: {
-			rules: [
-				{
-					test: /\.tsx?$/,
-					use: 'ts-loader',
-					exclude: /node_modules/
-				}
-			]
-		},
-		resolve: {
-			extensions: ['.tsx', '.ts', '.js']
-		},
-		externals: { fin: 'fin' },
-		output: {
-			filename: 'app-definition.bundle.js',
-			library: {
-				type: 'module'
-			},
-			path: path.resolve(__dirname, '..', 'public', 'js', 'modules', 'actions')
-		},
-		experiments: {
-			outputModule: true
-		}
-	},
-	{
 		entry: './client/src/modules/endpoints/example-connection-validation/index.ts',
 		devtool: 'inline-source-map',
 		module: {
