+++ how-to/customize-workspace/public/manifest.fin.json
@@ -834,11 +834,6 @@
 				},
 				{
 					"enabled": true,
-					"id": "app-definition",
-					"url": "http://localhost:8080/js/modules/actions/app-definition.bundle.js"
-				},
-				{
-					"enabled": true,
 					"id": "developer-actions",
 					"url": "http://localhost:8080/js/modules/actions/developer.bundle.js"
 				}
