+++ how-to/customize-workspace/public/second.manifest.fin.json
@@ -77,7 +77,6 @@
 						"checkLoginStatusInSeconds": 1,
 						"checkSessionValidityInSeconds": -1,
 						"customData": {
-							"userSessionId": "example-auth-data",
 							"users": [
 								{ "name": "Josh Smith (Developer)", "email": "josh@smith.com", "role": "developer" },
 								{ "name": "Sam Barns (Sales)", "email": "sam@barns.com", "role": "sales" }
@@ -93,7 +92,6 @@
 					"id": "auth-example-endpoint",
 					"url": "http://localhost:8080/js/modules/auth/example.bundle.js",
 					"data": {
-						"userSessionId": "example-auth-data",
 						"roleMapping": {
 							"developer": {
 								"excludeAppsWithTag": ["expero", "contact", "manager", "irs"],
