+++ how-to/customize-workspace/public/settings.json
@@ -879,11 +879,6 @@
 			},
 			{
 				"enabled": true,
-				"id": "app-definition",
-				"url": "http://localhost:8080/js/modules/actions/app-definition.bundle.js"
-			},
-			{
-				"enabled": true,
 				"id": "developer-actions",
 				"url": "http://localhost:8080/js/modules/actions/developer.bundle.js"
 			}
